oep-core/requirements.txt
aws-psycopg2
boto3
botocore
requests
